Particles seperated from testFlight
