+++ FlyingAces2/Assets/Scripts/Plane Scripts/PlaneParticleHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6e730275e3e293748a8aacc00ddaceb3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
