+++ FlyingAces2/Assets/Scripts/Plane Scripts/ThrowingChargeBarController.cs
@@ -22,19 +22,12 @@ public class ThrowingChargeBarController : MonoBehaviour
     private void OnEnable()
     {
         sliderValue = 0;
-        //StartCoroutine("UpdateSlider");
-    }
-
-    private void OnDisable()
-    {
-        //StopAllCoroutines();
     }
 
     private void FixedUpdate()
     {
         sliderValue += sliderIncrementValue;
         chargeBar.value = Mathf.PingPong(sliderValue * 3, 100);
-        //StartCoroutine("UpdateSlider");
     }
 
     private IEnumerator UpdateSlider()
